+++ src/knowledge_adapters/public_pdf/__init__.py
@@ -1,0 +1,2 @@
+"""Public PDF/report adapter."""
+